+++ test/widgets/flex_mount_track_test.dart
@@ -106,6 +106,7 @@ void main() {
         (tester) async {
       final mock = MockTracker();
       final client = await FlexTrackClient.create([mock]);
+      client.setGeneralConsent(true);
       addTearDown(() async {
         await client.dispose();
       });
@@ -133,6 +134,7 @@ void main() {
       final globalMock = await setupFlexTrackForTesting();
       final scopedMock = MockTracker();
       final scopedClient = await FlexTrackClient.create([scopedMock]);
+      scopedClient.setGeneralConsent(true);
       addTearDown(() async {
         await scopedClient.dispose();
       });